move dmg query command code into storage_query.go, set OmitPools flag in relevant requests and display one record line for led cmds
Required-githooks: true
