+++ src/control/cmd/dmg/storage_test.go
@@ -179,96 +179,10 @@ func TestStorageCommands(t *testing.T) {
 			nil,
 		},
 		{
-			"Set FAULTY device status (force)",
-			"storage set nvme-faulty --uuid 842c739b-86b5-462f-a7ba-b4a91b674f3d -f",
-			printRequest(t, &control.SmdQueryReq{
-				UUID:      "842c739b-86b5-462f-a7ba-b4a91b674f3d",
-				SetFaulty: true,
-			}),
-			nil,
-		},
-		{
-			"Set FAULTY device status (without force)",
-			"storage set nvme-faulty --uuid abcd",
-			"StorageSetFaulty",
-			errors.New("consent not given"),
-		},
-		{
-			"Set FAULTY device status (with > 1 host)",
-			"-l host-[1-2] storage set nvme-faulty -f --uuid 842c739b-86b5-462f-a7ba-b4a91b674f3d",
-			"StorageSetFaulty",
-			errors.New("> 1 host"),
-		},
-		{
-			"Set FAULTY device status without device specified",
-			"storage set nvme-faulty",
-			"StorageSetFaulty",
-			errors.New("the required flag `-u, --uuid' was not specified"),
-		},
-		{
 			"Nonexistent subcommand",
 			"storage quack",
 			"",
 			errors.New("Unknown command"),
-		},
-		{
-			"Reuse a FAULTY device",
-			"storage replace nvme --old-uuid 842c739b-86b5-462f-a7ba-b4a91b674f3d --new-uuid 842c739b-86b5-462f-a7ba-b4a91b674f3d",
-			printRequest(t, &control.SmdQueryReq{
-				UUID:        "842c739b-86b5-462f-a7ba-b4a91b674f3d",
-				ReplaceUUID: "842c739b-86b5-462f-a7ba-b4a91b674f3d",
-				NoReint:     false,
-			}),
-			nil,
-		},
-		{
-			"Replace an evicted device with a new device",
-			"storage replace nvme --old-uuid 842c739b-86b5-462f-a7ba-b4a91b674f3d --new-uuid 2ccb8afb-5d32-454e-86e3-762ec5dca7be",
-			printRequest(t, &control.SmdQueryReq{
-				UUID:        "842c739b-86b5-462f-a7ba-b4a91b674f3d",
-				ReplaceUUID: "2ccb8afb-5d32-454e-86e3-762ec5dca7be",
-				NoReint:     false,
-			}),
-			nil,
-		},
-		{
-			"Try to replace a device without a new device UUID specified",
-			"storage replace nvme --old-uuid 842c739b-86b5-462f-a7ba-b4a91b674f3d",
-			"StorageReplaceNvme",
-			errors.New("the required flag `--new-uuid' was not specified"),
-		},
-		{
-			"Identify device without device specified",
-			"storage led identify",
-			"",
-			errors.New("the required flag `-u, --uuid' was not specified"),
-		},
-		{
-			"Identify a device",
-			"storage led identify --uuid 842c739b-86b5-462f-a7ba-b4a91b674f3d",
-			printRequest(t, &control.SmdQueryReq{
-				UUID:     "842c739b-86b5-462f-a7ba-b4a91b674f3d",
-				Identify: true,
-			}),
-			nil,
-		},
-		{
-			"Clear identifying LED state on a VMD device",
-			"storage led clear --uuid 842c739b-86b5-462f-a7ba-b4a91b674f3d",
-			printRequest(t, &control.SmdQueryReq{
-				UUID:     "842c739b-86b5-462f-a7ba-b4a91b674f3d",
-				ResetLED: true,
-			}),
-			nil,
-		},
-		{
-			"Query LED state of a VMD device",
-			"storage led check --uuid 842c739b-86b5-462f-a7ba-b4a91b674f3d",
-			printRequest(t, &control.SmdQueryReq{
-				UUID:   "842c739b-86b5-462f-a7ba-b4a91b674f3d",
-				GetLED: true,
-			}),
-			nil,
 		},
 	})
 }
